+++ baybe/parameters.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 
 import logging
 from abc import ABC, abstractmethod
+from functools import cached_property, lru_cache
 
 from typing import ClassVar, Dict, List, Literal, Optional, Type, Union
 
@@ -18,6 +19,7 @@ from .utils import (
     df_drop_single_value_columns,
     df_drop_string_columns,
     df_uncorrelated_features,
+    HashableDict,
     is_valid_smiles,
     smiles_to_fp_features,
     smiles_to_mordred_features,
@@ -42,7 +44,15 @@ def _validate_value_list(lst: list, values: dict):
     return lst
 
 
-class Parameter(ABC, BaseModel, extra=Extra.forbid, arbitrary_types_allowed=True):
+class Parameter(
+    ABC,
+    BaseModel,
+    extra=Extra.forbid,
+    arbitrary_types_allowed=True,
+    keep_untouched=(
+        cached_property,
+    ),  # required due to: https://github.com/pydantic/pydantic/issues/1241
+):
     """
     Abstract base class for all parameters. Stores information about the
     type, range, constraints, etc. and handles in-range checks, transformations etc.
@@ -59,6 +69,12 @@ class Parameter(ABC, BaseModel, extra=Extra.forbid, arbitrary_types_allowed=True
     @classmethod
     def create(cls, config: dict) -> Parameter:
         """Creates a new parameter object matching the given specifications."""
+        return cls._create(HashableDict(config))
+
+    @classmethod
+    @lru_cache(maxsize=None)
+    def _create(cls, config: HashableDict) -> Parameter:
+        """Memory-cached parameter creation."""
         config = config.copy()
         param_type = config.pop("type")
         check_if_in(param_type, list(Parameter.SUBCLASSES.keys()))
@@ -79,7 +95,7 @@ class Parameter(ABC, BaseModel, extra=Extra.forbid, arbitrary_types_allowed=True
         #  member or a property, depending on the parameter type --> better solution?
         return item in self.values
 
-    @property
+    @cached_property
     @abstractmethod
     def comp_df(self) -> pd.DataFrame:
         """
@@ -131,7 +147,7 @@ class Categorical(Parameter):
     # validators
     _validated_values = validator("values", allow_reuse=True)(_validate_value_list)
 
-    @property
+    @cached_property
     def comp_df(self) -> pd.DataFrame:
         """
         See base class.
@@ -185,7 +201,7 @@ class NumericDiscrete(Parameter):
 
         return tolerance
 
-    @property
+    @cached_property
     def comp_df(self) -> pd.DataFrame:
         """
         See base class.
@@ -246,7 +262,7 @@ class GenericSubstance(Parameter):
         # for Python 3.7 or higher
         return list(self.data.keys())
 
-    @property
+    @cached_property
     def comp_df(self) -> pd.DataFrame:
         """
         See base class.
@@ -322,7 +338,7 @@ class Custom(Parameter):
         """
         return self.data.index.to_list()
 
-    @property
+    @cached_property
     def comp_df(self) -> pd.DataFrame:
         """
         See base class.